# backend_gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrepSense Gateway API",
...
    "version": "1.4.0",
    "contact": {
      "name": "PrepSense Development Team",
      "email": "dev@prepsense.com",
      "url": "https://github.com/dankimjw/PrepSense"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8002",
      "description": "Testing server"
    },
    {
      "url": "https://api.prepsense.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/nutrition/log-meal": {
      "post": {
        "tags": [
          "pantry",
          "nutrition"
        ],
        "summary": "Log Meal",
        "description": "Log a meal for dietary tracking.",
        "operationId": "log_meal_api_v1_nutrition_log_meal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogMealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogMealResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nutrition/gaps": {
      "get": {
        "tags": [
          "pantry",
          "nutrition"
        ],
        "summary": "Get Nutrient Gaps",
        "description": "Get nutrient gap analysis for a user.",
        "operationId": "get_nutrient_gaps_api_v1_nutrition_gaps_get",
        "parameters": [
          {
            "name": "analysis_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Date for analysis (defaults to today)",
              "title": "Analysis Date"
            },
            "description": "Date for analysis (defaults to today)"
          },
          {
            "name": "period_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of days to analyze (1 = single day, 7 = weekly)",
              "default": 1,
              "title": "Period Days"
            },
            "description": "Number of days to analyze (1 = single day, 7 = weekly)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNutrientGapsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nutrition/recipe-recommendations": {
      "post": {
        "tags": [
          "pantry",
          "nutrition"
        ],
        "summary": "Get Nutrient Aware Recipes",
        "description": "Get recipe recommendations with nutrient gap awareness.",
        "operationId": "get_nutrient_aware_recipes_api_v1_nutrition_recipe_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NutrientAwareRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nutrition/daily-intake/{target_date}": {
      "get": {
        "tags": [
          "pantry",
          "nutrition"
        ],
        "summary": "Get Daily Intake",
        "description": "Get daily dietary intake for a specific date.",
        "operationId": "get_daily_intake_api_v1_nutrition_daily_intake__target_date__get",
        "parameters": [
          {
            "name": "target_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Target Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nutrition/analyze-food": {
      "post": {
        "tags": [
          "pantry",
          "nutrition"
        ],
        "summary": "Analyze Food",
        "description": "Analyze nutritional content of a food item.",
        "operationId": "analyze_food_api_v1_nutrition_analyze_food_post",
        "parameters": [
          {
            "name": "food_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Food Name"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Quantity"
            }
          },
          {
            "name": "serving_unit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "serving",
              "title": "Serving Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/nutrition/nutrition-insights": {
      "get": {
        "tags": [
          "pantry",
          "nutrition"
        ],
        "summary": "Get Nutrition Insights",
        "description": "Get nutrition insights and trends for a user.",
        "operationId": "get_nutrition_insights_api_v1_nutrition_nutrition_insights_get",
        "parameters": [
          {
            "name": "period_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of days to analyze",
              "default": 7,
              "title": "Period Days"
            },
            "description": "Number of days to analyze"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetNutrientGapsResponse": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/NutrientGapAnalysis"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "priority_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Message"
          }
        },
        "type": "object",
        "required": [
          "analysis"
        ],
        "title": "GetNutrientGapsResponse",
        "description": "Response with nutrient gap analysis."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogMealRequest": {
        "properties": {
          "meal_type": {
            "$ref": "#/components/schemas/MealType"
          },
          "food_name": {
            "type": "string",
            "title": "Food Name"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity",
            "default": 1.0
          },
          "serving_unit": {
            "type": "string",
            "title": "Serving Unit",
            "default": "serving"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "meal_type",
          "food_name"
        ],
        "title": "LogMealRequest",
        "description": "Request to log a meal."
      },
      "LogMealResponse": {
        "properties": {
          "meal_id": {
            "type": "string",
            "title": "Meal Id"
          },
          "nutrients_added": {
            "$ref": "#/components/schemas/NutrientProfile"
          },
          "daily_total": {
            "$ref": "#/components/schemas/NutrientProfile"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "meal_id",
          "nutrients_added",
          "daily_total",
          "message"
        ],
        "title": "LogMealResponse",
        "description": "Response after logging a meal."
      },
      "MealType": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack",
          "beverage"
        ],
        "title": "MealType",
        "description": "Types of meals for dietary logging."
      },
...
      "NutrientAwareRecipeRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "include_nutrient_gaps": {
            "type": "boolean",
            "title": "Include Nutrient Gaps",
            "default": true
          },
          "focus_nutrients": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus Nutrients"
          },
          "meal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MealType"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_recipes": {
            "type": "integer",
            "title": "Max Recipes",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "NutrientAwareRecipeRequest",
        "description": "Request for nutrient-aware recipe recommendations."
      },
      "NutrientGap": {
        "properties": {
          "nutrient": {
            "type": "string",
            "title": "Nutrient"
          },
          "consumed": {
            "type": "number",
            "title": "Consumed"
          },
          "recommended": {
            "type": "number",
            "title": "Recommended"
          },
          "gap": {
            "type": "number",
            "title": "Gap"
          },
          "percentage_met": {
            "type": "number",
            "title": "Percentage Met"
          },
          "is_deficient": {
            "type": "boolean",
            "title": "Is Deficient"
          },
          "is_excessive": {
            "type": "boolean",
            "title": "Is Excessive"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "medium"
          }
        },
        "type": "object",
        "required": [
          "nutrient",
          "consumed",
          "recommended",
          "gap",
          "percentage_met",
          "is_deficient",
          "is_excessive"
        ],
        "title": "NutrientGap",
        "description": "Nutrient gap analysis for a user."
      },
      "NutrientGapAnalysis": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "analysis_date": {
            "type": "string",
            "format": "date",
            "title": "Analysis Date"
          },
          "period_days": {
            "type": "integer",
            "title": "Period Days",
            "default": 1
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/NutrientGap"
            },
            "type": "array",
            "title": "Gaps"
          },
          "priority_deficiencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Deficiencies"
          },
          "priority_excesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Priority Excesses"
          },
          "overall_score": {
            "type": "number",
            "title": "Overall Score",
            "default": 0.0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "analysis_date"
        ],
        "title": "NutrientGapAnalysis",
        "description": "Complete nutrient gap analysis for a user."
      },
      "NutrientProfile": {
        "properties": {
          "protein": {
            "type": "number",
            "title": "Protein",
            "default": 0.0
          },
          "carbohydrates": {
            "type": "number",
            "title": "Carbohydrates",
            "default": 0.0
          },
          "fiber": {
            "type": "number",
            "title": "Fiber",
            "default": 0.0
          },
          "total_fat": {
            "type": "number",
            "title": "Total Fat",
            "default": 0.0
          },
          "saturated_fat": {
            "type": "number",
            "title": "Saturated Fat",
            "default": 0.0
          },
          "trans_fat": {
            "type": "number",
            "title": "Trans Fat",
            "default": 0.0
          },
          "sugar": {
            "type": "number",
            "title": "Sugar",
            "default": 0.0
          },
          "vitamin_c": {
            "type": "number",
            "title": "Vitamin C",
            "default": 0.0
          },
          "vitamin_a": {
            "type": "number",
            "title": "Vitamin A",
            "default": 0.0
          },
          "vitamin_d": {
            "type": "number",
            "title": "Vitamin D",
            "default": 0.0
          },
          "vitamin_e": {
            "type": "number",
            "title": "Vitamin E",
            "default": 0.0
          },
          "vitamin_k": {
            "type": "number",
            "title": "Vitamin K",
            "default": 0.0
          },
          "thiamin": {
            "type": "number",
            "title": "Thiamin",
            "default": 0.0
          },
          "riboflavin": {
            "type": "number",
            "title": "Riboflavin",
            "default": 0.0
          },
          "niacin": {
            "type": "number",
            "title": "Niacin",
            "default": 0.0
          },
          "vitamin_b6": {
            "type": "number",
            "title": "Vitamin B6",
            "default": 0.0
          },
          "folate": {
            "type": "number",
            "title": "Folate",
            "default": 0.0
          },
          "vitamin_b12": {
            "type": "number",
            "title": "Vitamin B12",
            "default": 0.0
          },
          "calcium": {
            "type": "number",
            "title": "Calcium",
            "default": 0.0
          },
          "iron": {
            "type": "number",
            "title": "Iron",
            "default": 0.0
          },
          "magnesium": {
            "type": "number",
            "title": "Magnesium",
            "default": 0.0
          },
          "phosphorus": {
            "type": "number",
            "title": "Phosphorus",
            "default": 0.0
          },
          "potassium": {
            "type": "number",
            "title": "Potassium",
            "default": 0.0
          },
          "sodium": {
            "type": "number",
            "title": "Sodium",
            "default": 0.0
          },
          "zinc": {
            "type": "number",
            "title": "Zinc",
            "default": 0.0
          },
          "copper": {
            "type": "number",
            "title": "Copper",
            "default": 0.0
          },
          "selenium": {
            "type": "number",
            "title": "Selenium",
            "default": 0.0
          },
          "cholesterol": {
            "type": "number",
            "title": "Cholesterol",
            "default": 0.0
          },
          "omega_3": {
            "type": "number",
            "title": "Omega 3",
            "default": 0.0
          },
          "omega_6": {
            "type": "number",
            "title": "Omega 6",
            "default": 0.0
          },
          "calories": {
            "type": "number",
            "title": "Calories",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "NutrientProfile",
        "description": "Comprehensive nutrient profile for foods/recipes."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for external services"
      }
    },
...
  },
...
}